requirements.txt
openai==1.14.1
python-dotenv==1.0.0
